--- src/episodes/e1_lastwalk.py
@@ -1,110 +1,0 @@
-# -*- coding: utf-8 -*-
-"""Episode: last.do
-"""
-## import libraries
-import os
-import sys
-sys.path.append(os.path.join(os.path.dirname(__file__), '../..'))
-sys.path.append('storybuilder')
-from storybuilder.builder.world import World
-from storybuilder.builder.writer import Writer
-
-## alias
-W = Writer
-_ = W.getWho()
-
-
-## scenes
-def sc_lastwalk(w: World):
-    chi, ei = W(w.chisa), W(w.eisuke)
-    outside = W(w.outside)
-    return w.scene("最後の帰り路", "二人の歩幅はどんどん離れていく。そして",
-            w.comment("別れるところから始まる"),
-            w.symbol("　　　　十二月"),
-            chi.be(),
-            ei.be(),
-            chi.do("なんで、という言葉が白い息に変わってから消えた"),
-            ei.talk("だって、その、お前が最初に約束してくれって言っただろ？"),
-            chi.do("三十センチ上から、低い声が降ってくる",
-                "$meはじっと両手で鞄の把手を握り締め、まともに彼の顔を見上げられず、視線を彼の汚れの目立つスニーカーに向けたままで、",
-                "うん、と曖昧な返事をしてしまう"),
-            outside.look("マフラーを巻いたところから出た頬に当たる風が、冷たい",
-                "河川敷の土手は既に暗くて、いつも見かけるジョギングおじさんのニット帽が新しくなってるな、", "&"),
-            chi.think("なんて、どうでもいいことばかりが頭を巡る",
-                "周囲に同じ高校のブレザーは見かけない", "二人組の小学生がランドセルを背に歩いて通り過ぎていくけれど、",
-                "特に$meたちを気に掛ける様子もない"),
-            ei.talk("そうだろう？"),
-            chi.do("彼の声に、一度だけがんばって顔を上げてみる",
-                "両目に溜まった涙がすっと零れ、やっぱり$meは泣いてしまった",
-                "空は一部が茜色に塗られているだけで後は全部グレィの雲が蓋をしている",
-                "そこから雨が落ちてきそうだ"),
-            ei.talk("なあ"),
-            chi.do("自転車を脇に置いた彼はじっとこちらを見て何とか返事をもらおうとしているけれど、",
-                "$meはまだ、決断できずにいた"),
-            ei.talk("別れよう？"),
-            chi.explain("その瞬間というものは気づいたらもうそこに立ってしまっていて、後戻りができない",
-                "そういうものなんだなと、珍しくちょっと甘えたような言い方をした彼の声で理解した"),
-            chi.talk("あの"),
-            chi.talk("あの"),
-            chi.do("互いの声がハモる",
-                "どうしてこんな時だけ同じタイミングなんだろう"),
-            chi.talk("なんで、こうなっちゃったんだろうね"),
-            ei.talk("最初から合わなかっただけ、かもな"),
-            chi.think("彼の言う最初は、いつだろう",
-                "告白したときだろうか", "付き合い始めた時だろうか",
-                "関係がギクシャクしていると感じ始めた頃だろうか",
-                "$meは必死に彼を見ようとするけれど、目線を上げた時にはそっと彼の視線が明後日に向けられてしまう"),
-            chi.talk("ずっと、そう思っていたってこと？"),
-            chi.do("埋まらない空間を埋めるように質問を投げつける",
-                "思い返せばいつも会話はぎこちなく、どちらからともなく間をもたせようと質問を投げていた"),
-            ei.talk("そりゃ最初は好きだったよ",
-                "だから告白したんだ",
-                "けど、いつからかわかんねえけど、何か違うって思うようになって",
-                "最近は$lineも憂鬱だった"),
-            chi.think("それが初めて聞く、彼の本当の言葉だった"),
-            ei.talk("だって付き合って半年にもなるのに未だにキスの一つもしてないんだぜ？",
-                "それって付き合ってるって言えるのかよ？"),
-            chi.talk("え……"),
-            _.think("キスしたい、と思っていたことがショックだった",
-                "考えなかった訳じゃない",
-                "でもそれが決定的な不満になるとは思ってもなかった",
-                "ただ一緒にいて、帰る",
-                "学生だし、彼はバスケが忙しいし、そんなことなかなか考えられないのだと思い込んでいた"),
-            ei.talk("デートだって夏休みの一回こっきりだ",
-                "そりゃ$meだって忙しかったさ",
-                "けど、何だよ",
-                "付き合ってたらもっとこう、なんかあるだろ？",
-                "それともただの男友だちでしかなかったのかよ？"),
-            chi.talk("違う、けど"),
-            ei.talk("$meはお前の付き合うがよくわかんない",
-                "だからこうなるのも仕方ないんじゃないか？",
-                "わかり合えないなら、別れるしかない"),
-            chi.do("彼の真っ直ぐな目は$meを非難していた"),
-            ei.talk("じゃあ"),
-            ei.do("という、いつも橋で別れる時に口にするのと全く同じ言葉を吐き出して自転車に跨ると、",
-                    "背を向けてペダルを漕ぎ出した"),
-            chi.talk("$eisuke、君……"),
-            ei.do("呼びかけても何も答えずに行ってしまう"),
-            chi.do("左折して橋を渡り、そのままどんどん姿が小さくなる彼を$meはずっと見つめていた"),
-            chi.think("――どうしてこうなっちゃったんだろう"),
-            chi.do("それを考えながら、$meは空を見上げる",
-                    "世界を暗くする雲から落ちてきたのは小さな雪粒だった"),
-            # NOTE:
-            #   最後の帰り道は無言のまま、何故か歩幅が揃っている
-            #   どちらからも「別れよう」が出てくる
-            #   それを聞いて、ああ、やっぱり、と互いに思う
-            #   珍しく英輔は饒舌に語る
-            #   自分の知らない彼。でも最初の頃はこんなだった（あとで読者が気づく）
-            #   最初の約束、にフォーカスさせる
-            camera=w.chisa,
-            area=w.Tokyo,
-            stage=w.on_riverbed,
-            day=w.in_depart, time=w.at_evening,
-            )
-
-
-## episode
-def ep_lastwalk(w: World):
-    return w.episode("最後の帰り路", "いつの間にか歩幅は埋められなくなっていた",
-            sc_lastwalk(w),
-            )
