--- src/episodes/e2_change.py
@@ -1,145 +1,0 @@
-# -*- coding: utf-8 -*-
-"""Episode: change ours
-"""
-import os
-import sys
-sys.path.append(os.path.join(os.path.dirname(__file__), '../..'))
-sys.path.append('storybuilder')
-from storybuilder.builder.world import World
-from storybuilder.builder.writer import Writer
-
-## alias
-W = Writer
-_ = W.getWho()
-
-
-## scenes
-def sc_rainy(w: World):
-    chi, ei = W(w.chisa), W(w.eisuke)
-    outside = W(w.outside)
-    return w.scene("嵐の日に", "雨の日には相合い傘をするような関係だったのに",
-            w.symbol("　　　　十月"),
-            w.comment("季節は秋", "雨が多い"),
-            w.comment("呼び名は進展しているものの、彼の態度が、気持ちが離れつつある。ギクシャク感"),
-            outside.look("雨が降っている",
-                "冷たい雨だ",
-                "道路脇の側溝に汚くなった落ち葉が溜まっている",
-                "走り抜けていく車は水を跳ね、傘を差して歩く小学生の列に声を上げさせた"),
-            chi.be("そんな姿を眺めながら傘を手に、$meは待っている",
-                "郵便局の軒先から滴る雫がコンクリートの上に水たまりを作っていて、",
-                "それがどんどん大きくなっていくのを見ながら、待っている"),
-            chi.do("つい春先までは一人だったから誰かを待つなんてことはしなくても良かった",
-                "誰かを待つ",
-                "そのじりじりとした気持ちが恋の一部なんだと思っていたけれど、",
-                "何だろう、最近はその考えに違和感が生まれ始めていた"),
-            chi.do("鞄から携帯電話を取り出し、彼の連絡がないか確認する", "あった"),
-            _.do("先生に呼び出し受けて遅れると書かれていて、", "その言葉に少しほっとしたけれど、",
-                "携帯電話を握ったまま、もやもやとした気持ちと共にはっきりとしない空を見上げた"),
-            chi.think("秋季大会が近いからと彼のバスケ部は練習の日々で、",
-                "昨日はその様子を覗きに行ったりしたのだけれど、$meのことはちらりとも見てくれなかった",
-                "ただバスケットボールにだけ、仲間にだけ、視線を向けていた",
-                "その必死さが素敵だと思った", "でも寂しさも同時にあった"),
-            chi.do("まだ次の連絡はない",
-                "携帯電話に蓄積された彼との少ないメッセージのやり取りを振り返ると、", "&"),
-            chi.do("最初の頃は何気ないおはようやおやすみのメッセージが並んでいるだけ",
-                "彼はバスケの話で、シュートがよく決まったとか、練習試合でどうだったとか",
-                "それを$meが褒めたり、すごいと言ったり、",
-                "それでも時々は彼が好きそうな漫画やアニメの話題を振ってみる",
-                "でも「読んどく」「見てみるわ」が並んだ後には何もない"),
-            chi.think("だからいつも$meの思考は「どうして$meなの？」に辿り着く",
-                "彼から告白してきたのにこれでは何だか一方通行みたいで、",
-                "でもそういう真面目な話をできる雰囲気も今はまだない"),
-            chi.do("もう先に帰った方がいいかな、と思った時だった"),
-            ei.come("小雨を跳ね除けながら自転車を前傾姿勢で漕いできた彼が、目の前でブレーキを掛けて停まった"),
-            ei.talk("悪かった", "畑中説教がなげーんだ"),
-            ei.do("そこに浮かぶ笑顔は変わらない",
-                "ただヘルメットから雨粒が滴っている"),
-            chi.talk("今日は練習いいの？"),
-            chi.do("$meは慌てて傘を差し出す"),
-            ei.talk("休養日だってさ", "試合後だからそういう方針",
-                "練習しすぎても駄目だって今のコーチがさ"),
-            chi.talk("そっか"),
-            chi.do("傘の先から落ちた雨が、彼の肩を濡らしていた", "思い切り高くしてみるけれど身長差はどうしようもないみたいだ",
-                "その差は永遠に埋まらない気がして何だか情けなくなる"),
-            chi.talk("ごめん"),
-            ei.talk("ああ、$meが持つわ"),
-            ei.do("彼は自転車から降りると傘を奪うように受け取り、代わりに持ってくれた",
-                "&"),
-            chi.do("でも見上げる先にある傘の赤い布が随分遠い",
-                "$meは一歩、彼に近づいたけれど、それでも背中か、肩か、今度は$Sに雫がかかる", "どこに立っても$meの一部があぶれてしまった"),
-            ei.talk("傘ぱくってくればよかったな", "悪い"),
-            chi.talk("盗むのは駄目だよ", "$meがもっと大きな傘持ってくればよかっただけ", "ごめん"),
-            _.think("最近ごめん、と口癖のように謝っている", "彼もよく悪いって言うばかりで、",
-                "何だろう、このギクシャクした居心地の悪さは",
-                "楽しい、も、嬉しい、も、最近あまり見かけない"),
-            chi.talk("なんか、ごめん"),
-            ei.talk("いいよ", "それより帰ろうぜ？"),
-            chi.talk("うん"),
-            chi.do("頷いた途端に目から雫が落下した"),
-            ei.talk("おいおい", "なんでそう泣くんだよ", "$me、また何か言ったか？"),
-            chi.talk("ううん。傘が、小さかったから"),
-            chi.do("泣きたい訳じゃないけれど、雨に濡れただけなんて誤魔化しを口にすることもできないくらい、",
-                    "精神が疲れているのが分かった"),
-            ei.talk("ああ、もう分かったよ"),
-            chi.do("そう言うと彼は$meに傘を押し付けて、先に歩いて行ってしまう",
-                    "自転車も彼も、どんどん濡れていく",
-                    "ブレザーの紺色が濃くなり、ズボンには滝のようなシミができていた"),
-            chi.talk("待って！"),
-            chi.do("$meは彼を追いかける"),
-            chi.talk("濡れてる！"),
-            ei.talk("別にいいよ、これくらい", "$chisaが泣くよりマシだから"),
-            chi.do("以前ならその言葉と笑顔に救われていた",
-                    "でも今は、付け合せの野菜みたいにどこかしんなりとして、味気ない"),
-            chi.talk("ごめん"),
-            chi.do("そう言って付いていく"),
-            outside.look("交差点の信号は赤くぼやけて、二人の足を止めようとするけれど、", "&"),
-            chi.do("彼は構わずに自転車に跨ると、呻き声を上げて渡って行ってしまった"),
-            chi.talk("$ln_eisuke君！"),
-            chi.do("$meはその後を追いかけられず、濡れた横断歩道を前に、走ることを諦めた"),
-            # NOTE:
-            #   雨の中、秋になっている
-            #   傘を使って互いに近寄れない、心の距離が離れていっていることを見せる
-            camera=w.chisa,
-            stage=w.on_crossroad,
-            day=w.in_disquiet1, time=w.at_evening,
-            )
-
-def sc_apart(w: World):
-    chi, ei = W(w.chisa), W(w.eisuke)
-    outside = W(w.outside)
-    return w.scene("離れる距離", "立ち止まらなくなった彼",
-            chi.talk("ねえ"),
-            chi.do("彼は振り返らないのを"),
-            chi.talk("ねえってば"),
-            ei.talk("ん？"),
-            chi.talk("最近スマホばかり見てる"),
-            ei.talk("そうか？"),
-            chi.talk("やっぱ変。なんかあった？"),
-            ei.talk("なんもないよ"),
-            chi.talk("キスの一つもしてないから？"),
-            ei.talk("ちげーよ。そういうんじゃない"),
-            chi.talk("だったら何？"),
-            ei.do("じっと"),
-            chi.talk("もうちょっと、話してほしい"),
-            ei.talk("何話せばいいんだよ"),
-            chi.talk("なんでもいい。ゲームでもバスケでも何でも"),
-            ei.talk("お前としても楽しくないだろ？"),
-            chi.talk("会話がしたいんじゃない"),
-            ei.talk("わかんねーな"),
-            chi.think("彼の歩く速度が早いと"),
-            # NOTE:
-            #   別日。秋風が冷たくて。コート
-            #   最初から距離が離れている
-            #   少し大きな声で呼び止めて、バスケにたとえられる
-            #   不和。話が噛み合わない
-            #   なんで、をスルーされる
-            stage=w.on_riverbed,
-            day=w.in_disquiet2, time=w.at_evening,
-            ).omit()
-
-## episode
-def ep_change_our(w: World):
-    return w.episode("変わる二人", "少しずつ変わっていく二人",
-            sc_rainy(w),
-            sc_apart(w),
-            )
