--- src/episodes/e3_enjoy.py
@@ -1,138 +1,0 @@
-# -*- coding: utf-8 -*-
-"""Episode: enjoy
-"""
-import os
-import sys
-sys.path.append(os.path.join(os.path.dirname(__file__), '../..'))
-sys.path.append('storybuilder')
-from storybuilder.builder.world import World
-from storybuilder.builder.writer import Writer
-
-## alias
-W = Writer
-_ = W.getWho()
-
-
-## scenes
-def sc_handdistance(w: World):
-    chi, ei = W(w.chisa), W(w.eisuke)
-    outside = W(w.outside)
-    return w.scene("手の届く距離で", "まだ二人は手の届く距離を歩いていた",
-            w.comment("少し季節が動く"),
-            w.symbol("　　　　七月"),
-            chi.come("#歩いている"),
-            ei.come("#歩いている"),
-            outside.look("夏服の白シャツは半袖の袖口が肌に張り付いて、やや心地悪い",
-                "その肌の上を生温い風が抜けていく",
-                "河原では子どもたちが水浴びをしてきゃっきゃと声を上げていて、自転車を引いて歩いている彼が「いいな」と笑う", "&"),
-            chi.do("それを追いかけるように隣で$meも「うん」と相槌する"),
-            _.do("この一月半ばかりの下校時の日常だった"),
-            chi.do("鞄を彼の自転車の前カゴに放り込み、両手をぶらぶらと手持ち無沙汰で歩く",
-                "彼の部活がない日は揃って下校できるからいつもの三倍は空気が美味しい"),
-            ei.talk("$ln_chisa？"),
-            chi.talk("あ、うん", "ごめん"),
-            chi.do("気づくと彼が二メートルも先に進んでいる",
-                "$meは慌てて追いつく",
-                "でもその追いつくまで待っていてくれるのが嬉しい",
-                "恋は沢山の嬉しいが待っていると知れたのも、彼のお陰だ"),
-            chi.talk("県大会、残念だったね"),
-            ei.talk("まあ、な"),
-            chi.do("ずっと言えなかった言葉を口にできてほっとする",
-                "けれど彼は雲のほとんど浮いてない空を見上げて「うん」と溜息を吐く",
-                "その吐息の中にどんな気持ちが含まれているのかは、決して話してくれない"),
-            ei.talk("それより試験の方が頭痛い"),
-            chi.talk("$ln_eisuke君の場合はそうかも知れないね"),
-            ei.talk("何だよ", "これでも国語はちっとはマシなんだぞ？"),
-            chi.think("前に最高得点六十八点だったと自慢そうに語ったことを思い出し「そうだね」と笑った"),
-            ei.do("笑う","&"),
-            chi.think("それがずっと欲しかったものだった",
-                "彼の笑顔", "日陰にいるみたいな自分を照らしてくれそうな、優しい、歯並びが少し悪い、でもとても好きな笑顔がすぐそこにある"),
-            chi.do("自転車のハンドルを握る彼の手はがっちりとして節がしっかりしている",
-                "それをじっと見ていると彼が「ん？」と目線を向ける"),
-            chi.talk("ううん。なんでもない"),
-            chi.think("先に歩き出して$meはバスケをしている時のパスを受けて高くボールを掲げた時の、敵と味方を見極めているような横顔を思い出した恥ずかしさをそっと隠す"),
-            ei.talk("ああ、あれか", "また$meが先に行っちゃったから"),
-            chi.talk("ううん、それは別にいいって言ったよ？", "$meが合わせればいいんだし"),
-            chi.do("二人の歩幅は違う", "気にせずに歩き出すとすぐに大差がついて$meは置いてけぼりだけれど、",
-                "彼に付いていくと決めたから文句は言わずに歩く", "最初の頃に何度も立ち止まって溜息がちに見るものだから約束したのだ――$meががんばって追いつくから"),
-            ei.talk("いや、でもさ、どう考えても$ln_chisaに無理させてるって最近思って",
-                "バスケやってるとさ、ほんとチームって大事だなって思うんだ"),
-            chi.talk("またバスケの話？"),
-            chi.think("彼の話の大半はバスケだった", "漫画も、見るテレビも、ゲームも、何もかもバスケ",
-                "友だちも男ばかりで、遊びに行くよりもとにかくバスケの練習がしたいと$meにまで言うような、",
-                "一人のバスケ好きな男子高校生が彼だった"),
-            ei.talk("いや、バスケの話ってか、バスケの話じゃないけど、やっぱバスケの話か"),
-            chi.do("考え込むように顎に右手を当てていたけれど、そのまま歩き出し、話を続けた"),
-            ei.talk(
-                "一人上手い奴がいてもさ、バスケって一人じゃできないんだ",
-                "ボールを運んでシュートが決まっても、すぐにそんな点は取り返されちまう",
-                "結局攻めるにも守るにも周りの奴らとの協力が必要になるんだ",
-                "だから二人でも三人でもいいけど、チームっていうのがとにかく大事でさ、",
-                "それって自分のことだけじゃなくて相手がどう動くか、どうパスするのか、どこにポジショニングするのかとか、",
-                "注意深く見て理解してなきゃいけない",
-                "$meと$ln_chisaもそんな関係を目指していけばいいんじゃないかなって思うんだけど……分かる？"),
-            chi.do("バスケのことになると息の継ぎ間がないくらいに一気呵成にまくしたてて喋る",
-                    "そもそもバスケットボール自体あまり詳しくないし、時々分からない用語が出るけれどそれをいちいち尋ねてもいられないしで、",
-                    "いつも彼が伝えたいことは半分も理解してあげられない"),
-            chi.talk("$meたちはチームってことかな？　カップルじゃなくて"),
-            ei.talk("英語はよく分かんねえけど、一緒に帰れる時はなるべく一緒に帰ろうって、そういうことかな"),
-            chi.talk("うん、わかった。ありがとうね"),
-            chi.do("笑顔を作る",
-                    "彼のそれとは比べ物にならない、作り物でしかない",
-                    "それでも彼は満足して「ああ」とにっこりしてくれるから、たぶんこれで間違っていない"),
-            outside.look("河川敷の公園ではサッカーをしている小学生たちが楽しそうな声を上げている",
-                "一人の子が蹴ったボールが川に落ちて、それでも気にせずに別の子が入っていって拾ってあげた",
-                "&"),
-            chi.do("自分ならきっと黙って見ている側だろう",
-                    "けど彼はああやって拾ってあげる側だ",
-                    "そんな違いをよく感じる",
-                    "それは遠くから見ているだけでははっきり分からなかった、$meと彼との距離感なんだと思った"),
-            chi.do("でもその違う他人が互いを補い合うように歩み寄って、チームになる",
-                "きっとそういうことなんだ", "それが$meたちの恋だと、彼は言いたいのかも知れない",
-                "そんな考えに至り、頬が緩む"),
-            chi.talk("今度の休み、買い物とか、行く？"),
-            ei.talk("ほんと！？", "あ、けど次の休みは高村たちと試験勉強で集まるんだったわ。悪い"),
-            chi.talk("ううん", "それじゃあ試験終わったら、また考えよう"),
-            ei.talk("そうだな"),
-            chi.do("彼の笑顔に、胸の閊えが取れた思いだ"),
-            chi.do("でも顔を上げると、お別れポイントが近づいていた"),
-            outside.look("鉄骨で組まれた橋が向こう岸まで伸びている",
-                    "その橋が自転車通学と徒歩通学の境界線で、", "&"),
-            chi.explain("$meと彼を分けているものでもある"),
-            ei.talk("もうすぐだな"),
-            chi.talk("うん"),
-            chi.do("いつも繰り返される、確認のようなそれを、彼はどんな思いで口にしているのだろう", "&"),
-            _.think("ちら、と表情を伺う",
-                    "ちょっと落ち着きがないように見えるのは別れを惜しんでくれているからだろうか"),
-            ei.talk("な、ちょっと聞きたいんだが"),
-            chi.do("彼は歩くのをやめて$meの方を振り返る",
-                    "その目が真剣な様子で、思わず何が来るのだろうと身構えてしまうけれど、"),
-            ei.talk("好きな食べ物ってあるか"),
-            chi.do("なんてことのない質問だった"),
-            chi.talk("好きは特別ないけど嫌いなら結構ある"),
-            ei.talk("$meはグリンピースが駄目でさ、カレーとかピラフとかオムライスとかさ"),
-            chi.talk("うんうん。わかる"),
-            chi.do("それで彼は満足したのか、何度も頷いてから自転車に跨る"),
-            chi.think("思わず「待って」と口にしたくなった", "もう少しだけ一緒にいないかと提案したい", "でも彼はあの笑顔を$meに向けて、"),
-            ei.talk("じゃあ"),
-            ei.do("軽く手を振り、ペダルを漕いで行ってしまう"),
-            chi.talk("うん", "また明日……ばいばい"),
-            chi.do("彼の自転車が橋を渡っていく",
-                    "距離はどんどん離れ、ついさっきまで感じていた肌の熱はいつの間にかひんやりとした心地になり、胸の中央をすっと風が抜けていく",
-                    "それでも$meはいつも彼が橋を渡り切ってしまうまでじっと見つめ、ある期待をする",
-                    "でも今日も彼は振り返らない",
-                    "まだ心の距離が遠い気がした"),
-            # NOTE:
-            #   手を繋げないまま、歩くが、互いに目を合わせたり、彼の話に適当に頷くのすら楽しい日々
-            #   好きな食べ物の話なんかでも楽しい
-            #   けれど尽く「合わない」
-            camera=w.chisa,
-            stage=w.on_riverbed,
-            day=w.in_enjoy1, time=w.at_afternoon,
-            )
-
-## episode
-def ep_enjoytime(w: World):
-    return w.episode("楽しかった頃", "まだあの頃は一緒に帰る道が楽しいばかりだった",
-            sc_handdistance(w),
-            )
