--- src/episodes/e4_firststep.py
@@ -1,235 +1,0 @@
-# -*- coding: utf-8 -*-
-"""Episode: first step
-"""
-import os
-import sys
-sys.path.append(os.path.join(os.path.dirname(__file__), '../..'))
-sys.path.append('storybuilder')
-from storybuilder.builder.world import World
-from storybuilder.builder.writer import Writer
-
-
-## alias
-W = Writer
-_ = W.getWho()
-
-## scenes
-def sc_confession(w: World):
-    chi, ei = W(w.chisa), W(w.eisuke)
-    inside, outside = W(w.inside), W(w.outside)
-    return w.scene("告白の鼓動", "最初の瞬間から、その不安は過っていた",
-            w.comment("告白される", "場所は学校を少し離れたところがいい"),
-            w.comment("季節は春から初夏くらいが良い。最終的に冬に別れるから"),
-            w.comment("$eisukeだけは自転車を引っ張っている。通学が異なる"),
-            w.symbol("　　　　五月"),
-            w.comment("身長x0.45が目安"),
-            chi.come("身長百五十一センチの歩幅は約六十八センチらしい", "&"),
-            chi.do("その幅を重ねて$meは高校を出て一キロくらいの場所にある横断歩道に近づく",
-                "温かい風が小学校のフェンス越しの葉桜を揺らしていくのを横目に歩いていく$meを、",
-                "黄色いカバーを付けたランドセルを背負った小さな子たちが集団で追い抜いていく",
-                "あの頃からずっと、大きな声で子ども特有の高い音域で喚くように喋り合う同級生たちを、",
-                "少し離れて羨ましさを噛み殺しながら見ていた",
-                "そんな$meは今でもこうして一人で、家路を歩いている"),
-            chi.do("横断歩道の信号は赤へと変わり、$meと小学生の集団を分断した",
-                "メロディは止まり、デイサービスの送迎車が慌てて黄色信号を渡っていく"),
-            chi.look("鞄をスカートの前にして立ち止まると、紺色の野暮ったいブレザーが陽射しに温められる","&"),
-            _.think("衣替えはまだ先だからと体操着に着替えて「知るかよ」と友だちと笑っていた同級生の角刈りの男子の笑顔を思い出し、",
-                "$meも前の二つボタンを外してみた", "汗ばむ白シャツにいくらか外気が入り込んで、気持ち、涼しくなった"),
-            _.think("学校の中ではまともに彼を見ることも、ちょっと真似をしてみることでさえも、全然できない",
-                "それでもこうやってあの笑顔を思い出すだけで元気が湧いた", "$meにとって、彼は、やや息苦しさを感じる学生生活の光明だった"),
-            chi.hear("体育館を覗きに行くといつも大きく荒い息で「パース！」と叫んでいる",
-                "重そうな体を揺すって走ってはあの大きな手でボールを掴み、がむしゃらにゴールに向かっていく",
-                "こうして目を閉じて耳を澄ますだけで、頭の中で彼の息遣いが再生されるのだ"),
-            chi.think("そう思っていた"),
-            ei.come("#そこに$Sが立っていた"),
-            ei.talk("あ、あの！"),
-            chi.do("急ブレーキの耳障りな音に続いた大きな声に振り返るとそこに、$meより三十センチも大きな体操着姿の男性が肩を揺らして立っていた", "自転車から降りて、慌ててスタンドを立てる",
-                "彼は一度下を向き、それから顔を上げてもう一度「あの」と息混じりの声で$meに話しかける",
-                "それが本当に$meなのか分からなくて左右を見て、他に人がいないのを確認したけれど、彼は「$ln_chisaさん」と呼ぶから思わず「はい」と返事をしてしまった"),
-            chi.do("真正面から顔を見たことなんてないから彼の太くて力強い眉が困ったように曲がり、",
-                "それでも真剣な眼差しが自分に向けられていて、一秒くらいは我慢したけれど、思わず顔を背けてしまう",
-                "明らかに気まずい"),
-            ei.talk("いや、その、別に怖い話じゃないんだ"),
-            chi.talk("う、うん"),
-            chi.do("ぎゅっと掴まれたような喉から無理矢理に出した声が、裏返る"),
-            ei.talk("あの、今さ、付き合ってる人とか、その、彼氏とか、いる？"),
-            chi.talk("え？"),
-            _.do("声にならない掠れた呟きを、$meは何度も自分の胸の内で繰り返す",
-                "彼は短い髪の毛をガリガリと掻き毟り、口なのか鼻からなのか分からない荒い息遣いで答えを待っている",
-                "そのハァハァスゥ、という彼の息のリズムが$meの鼓動を大きく狂わせ、ぎゅっと胸の前に右手を持っていく",
-                "苦しい、という気持ちを辛うじて「ううん」という返事に変える"),
-            ei.talk("え？"),
-            chi.talk("い、いません"),
-            chi.do("はっきりと否定することに唇が震える"),
-            ei.talk("じゃあ、好きな人とかは？"),
-            chi.talk("えっ"),
-            ei.talk("あ、ごめ", "今の無し", "ちょっと待って"),
-            chi.do("答える暇もなく彼が慌ててその大きな逞しい手を広げて質問をかき消そうとするのが何だかおかしくて、",
-                "小さく吹き出した"),
-            ei.talk("あ、だから、そうじゃなくて"),
-            ei.do("彼は咳払いをして深呼吸で整えると、いつも友だちと話しているのとは全然違う低く、けれどしっかりと伝わる声でこう言った"),
-            ei.talk("$meと、付き合って下さい", "お願いします"),
-            ei.do("右手を出し、思い切り頭を下げる"),
-            chi.do("$meは慌ててとりあえず周囲を見た", "&"),
-            outside.look("車は信号の切り替わりで止まり、横断歩道は青に変わり、待っていたおばさんが歩き出す",
-                    "新しい小学生の集団の中の一人の女子は向かい合って何だか気まずそうな高校生二人を見て、何かを察しているのかも知れない",
-                    "自転車通学の知らない生徒の男子が、目線だけ向けて$meたちの脇を抜けていった"),
-            chi.do("顔が熱い", "頬が火照る",
-                    "彼の顔なんてもう見れず、それでもじっと右手を出したまま待っているのだと分かる"),
-            chi.do("どうすればいい？", "どう答えればいい？"),
-            _.do("その自問が頭の中で繰り返される",
-                    "まるで左脳と右脳が喧嘩を始めたみたいに思考の箱がいっぱいになって、溢れて、",
-                    "$meは何も言えないまま鼻水をすすり始める"),
-            ei.talk("あの、$ln_chisaさん？"),
-            chi.hear("横断歩道のメロディは音が半分外れたような調子でずっと鳴り響いている"),
-            chi.talk("なんで、$me？"),
-            chi.do("やっと絞り出したのはそんな問いかけで、", "彼がどんな表情をしているかなんて見られずに$meは胸の右手を再び鞄の握りに戻すと、",
-                    "俯いたままの瞳から涙を落とした"),
-            ei.talk("なんでって、その、好きだけじゃ駄目、かな"),
-            chi.hear("好き、という言葉の響きは甘酸っぱい", "でもそれは砂糖なのか毒なのか、とても判別できそうにない",
-                    "指先からじわりと痺れが広がって、体温が行方不明だ"),
-            ei.talk("好きだけじゃ、駄目かよ？"),
-            chi.hear("声が強くなる",
-                    "周りの人にもきっと聞こえている"),
-            chi.think("逃げ出したい",
-                    "それなのに「駄目か？」と彼の声が追い打ちをする",
-                    "$meの足にそれぞれ十キロくらいの重石を乗せる"),
-            ei.talk("……泣いて、る？"),
-            chi.do("声は出ず、ただ頷くだけだ"),
-            ei.talk("じゃあ駄目か？"),
-            ei.do("諦めたように呟いて、彼の温度が遠ざかる"),
-            chi.talk("駄目、じゃないよ"),
-            chi.do("顔を上げた", "彼は自転車に跨がり、行ってしまおうとしている",
-                    "でも「じゃないよ」ともう一度言うと顔をこちらに向けてくれた"),
-            ei.talk("あの、泣かせるつもりじゃなかった",
-                    "てか、泣くと思ってなかった",
-                    "突然こんなこと言って、ごめんかった"),
-            chi.talk("ごめん、くない", "ごめんくないよ"),
-            ei.talk("え？"),
-            chi.talk("嬉しいって言ったの", "嬉しいんだよ？"),
-            chi.do("彼は困惑しつつも「じゃあ」と$meの顔を覗き込む"),
-            chi.do("それに小さく頷くと、やっと彼の表情が明るくなった",
-                    "気持ちが通じた、と思える瞬間だった"),
-            ei.talk("そっか。そっかよ"),
-            chi.do("彼は自転車を降りて何度も「そっか」と呟く", "声が上擦って、でもとても嬉しそうで何度か言っているうちに「やった」「やった！」とガッツポーズまで始める"),
-            chi.talk("$ln_eisuke君、ちょっと"),
-            outside.look("信号はまた赤に変わり、目の前を車が走り抜けていった"),
-            chi.talk("危ないよ？"),
-            ei.talk("大丈夫大丈夫", "今の$meは無敵だから"),
-            ei.do("笑顔だ", "&"),
-            chi.do("彼の太陽みたいな笑顔が、すぐそこにあって、自分へと向けられている",
-                    "ずっと欲しかった、優しくて温かいそれが、自分のものになった瞬間が、この時だった"),
-            ei.talk("えっと、これから、宜しくお願いします"),
-            chi.talk("こ、こちらこそ", "お願いします"),
-            w.comment("ある橋まで、で分かれていることを示す"),
-            # NOTE:
-            #   告白の返事をした後の、ぎこちなさから開始
-            #   一緒に帰ろうか、のところ
-            #   まだ春の残り香が街にある
-            #   爽やかな空とかの情景
-            camera=w.chisa,
-            stage=w.on_crossroad,
-            day=w.in_first, time=w.at_afternoon,
-            )
-
-def sc_samestep(w: World):
-    chi, ei = Writer(w.chisa), Writer(w.eisuke)
-    inside, outside = W(w.inside), W(w.outside)
-    return w.scene("歩幅を合わせて",
-            w.comment("季節は２学期が始まりで、秋。まだ残暑"),
-            w.comment("告白を終えて、一緒に初めて帰った時の歩幅。合わせようとしていた。どちらからともなく"),
-            w.br(),
-            chi.come("#歩いている"),
-            ei.come("#前を歩いている"),
-            outside.look("夕陽が河川敷の片方の土手を大きく陰にしてしまう",
-                "川は水量が少なくてぽつりぽつりと離れ小島みたいな岩がいくつか見えて、","&"),
-            chi.do("そこを桜の花びらが流れていく", "前を向くと五メートルくらい彼との距離が離れてしまっていた"),
-            chi.do("待って、とも言えず、小さな歩幅で懸命に追いつく",
-                "彼は自転車を押しているだけなのに随分と早くて、のんびりと歩く$meとはすぐに差がついてしまった"),
-            ei.talk("あ、悪い"),
-            chi.talk("ううん", "気にしないで"),
-            chi.do("彼は肩越しに振り返り、立ち止まって$meを待ってくれる",
-                "百八十の彼の歩幅は約八十一センチ、一歩で十センチ以上も$meを置いていってしまう",
-                "だから$meが早足になって追いかけるしかなくて、",
-                "一緒に帰ろうと言ってくれた彼は面倒な女だとでも思っているかも知れない",
-                "そんなことを考えると、目に涙が溜まる"),
-            chi.talk("あの、さ"),
-            stage=w.on_riverbed,
-            day=w.in_first, time=w.at_evening,
-            )
-
-def sc_samestep2(w: World):
-    chi, ei = Writer(w.chisa), Writer(w.eisuke)
-    inside, outside = W(w.inside), W(w.outside)
-    return w.scene("歩幅を合わせて2",
-            ei.talk("ん？"),
-            chi.hear("彼の低い声が、響く"),
-            chi.do("何か話していないと泣いてしまいそうで、けれど話題を探してもまず言葉が浮かんでこなくて、",
-                "目線は下を向き、歩く速度はどんどん遅くなって立ち止まってしまう", "だから"),
-            chi.talk("何でもない"),
-            chi.do("そう言って諦めた"),
-            ei.talk("$ln_chisaさんはさ、やっぱ本とか読んだりするタイプなの？"),
-            chi.do("けれどそんな$meに彼は優しく話しかける"),
-            ei.talk("ほら、いつも教室では大人しくしてるグループじゃん",
-                "$ln_miyoさんとかと一緒で"),
-            chi.talk("う、うん"),
-            chi.do("唯一話しかけてくれるのが前の席の$full_miyoだった",
-                "ただ友だちと呼ぶまでの関係性はまだ築けていない"),
-            ei.talk("$ln_chisaさんはさ、他の女子みたいにうるさくない人だから、やっぱ本かなって思って"),
-            chi.think("質問の意図が呑み込めなかった",
-                "それでも彼の言葉から何を言いたいのか、その気持ちをがんばって汲み取ろうと次の言葉を待った"),
-            ei.talk("$meは教科書もろくに読まないんだけど、漫画だと読む",
-                "って言ってもやっぱバスケ漫画が嬉しいけど",
-                "ただそれだけじゃバカになるからって兄貴がやたらと日本の歴史とか三国志とか貸してくれるんだ",
-                "知ってるかな、三国志"),
-            chi.talk("うん", "有名な中国の歴史ものだよ"),
-            ei.talk("兄貴はやたら三国志が好きみたいでゲームも三国志、漫画も三国志、小説も三国志、おまけに人生も三国志から学べって",
-                "将来三国志学者にでもなればいいんじゃねえかなって本気で思ってる"),
-            chi.do("そう思わない？　という、いつも友だちにして見せるみたいな悪戯坊主の笑みが$meに向けられる", "&"),
-            chi.do("たぶん話題を作ろうとしてくれているのだろう",
-                "その心遣いが少し息苦しい", "けどその苦しさが恋なのかも知れないと、$meは思っていた"),
-            chi.talk("そうだね"),
-            chi.do("$meの返事を待ってから、彼はゆっくり歩き出す",
-                "自転車の車輪のカラカラと回る音が少しずつ自分から離れていく",
-                "あ、と思ってそれに必死で追いついて、でもまた離れて、を繰り返す"),
-            outside.look("いつの間にか河川敷の路を歩く人影は、$meと彼、二つになっていた",
-                "その間を風が抜けていく"),
-            chi.talk("少し涼しくなってきたね"),
-            ei.talk("そうか？"),
-            chi.look("彼は体操着で、もう一人だけ先に夏にいるみたい",
-                    "対して自分はまだ冬服のブレザーが重い", "ボタンを外しているくらいでは、彼に追いつけない"),
-            chi.talk("$ln_eisuke君は、いつも涼しそう"),
-            ei.talk("ああ、コレな", "女子はジャージや体操着って訳にはいかないもんな"),
-            chi.do("彼はシャツを引っ張って$meを見る",
-                    "まださっき告白されたところなのにもうこんなに近くで自分を見てくる",
-                    "好きだから？　恥ずかしくないのかな",
-                    "分からない",
-                    "その照れ臭さを隠すように下を向いて早足になるけれど、彼はすぐに追いついて、行き過ぎたとまた立ち止まる",
-                    "顔を上げるとそこに、彼の笑顔が待っていた"),
-            chi.talk("ねえ、あの、さ"),
-            # NOTE:
-            #   歩幅を合わせようと、千紗が急ぎ足で追いつこうとするところ
-            #   そこで背中を見つめて、思うこと
-            ).omit()
-
-def sc_samestep3(w: World):
-    chi, ei = Writer(w.chisa), Writer(w.eisuke)
-    inside, outside = W(w.inside), W(w.outside)
-    return w.scene("歩幅を合わせて3",
-            ei.talk("何？"),
-            chi.talk("もしどっちかが駄目ってなったら、その時は後腐れなく、別れるって約束、して下さい"),
-            chi.do("彼は一瞬何を言われたのか分からなかったみたいだったけれど、一分ほど考え込んで、"),
-            ei.talk("分かった"),
-            outside.look("夕日が落ちてくる",
-                "二人の影が長く伸び、歩き出した時に重なっていたそれは、ゆっくりと離れていった"),
-            w.symbol("（了）"),
-            )
-
-## episode
-def ep_firststep(w: World):
-    return w.episode("最初の帰り路", "一番最初から、その予感があったことを思い出す",
-            sc_confession(w),
-            sc_samestep(w),
-            sc_samestep2(w),
-            sc_samestep3(w),
-            )
