+++ src/scenes/Riverbed.py
@@ -1,0 +1,530 @@
+# -*- coding: utf-8 -*-
+"""Stage: Riverbed
+"""
+## import libraries
+import os
+import sys
+sys.path.append(os.path.join(os.path.dirname(__file__), '../..'))
+sys.path.append('storybuilder')
+from storybuilder.builder.world import World
+
+
+## scenes
+def sc_lastwalk(w: World):
+    chi, ei = w.get("chisa"), w.get("eisuke")
+    return w.scene("最後の帰り路", "二人の歩幅はどんどん離れていく。そして",
+            w.comment("別れるところから始まる"),
+            w.symbol("　　　　十二月"),
+            chi.be(),
+            ei.be(),
+            chi.do("なんで、という言葉が白い息に変わってから消えた"),
+            ei.talk("だって、その、お前が最初に約束してくれって言っただろ？"),
+            chi.do("三十センチ上から、低い声が降ってくる",
+                "$meはじっと両手で鞄の把手を握り締め、まともに彼の顔を見上げられず、視線を彼の汚れの目立つスニーカーに向けたままで、",
+                "うん、と曖昧な返事をしてしまう"),
+            chi.look("マフラーを巻いたところから出た頬に当たる風が、冷たい",
+                "河川敷の土手は既に暗くて、いつも見かけるジョギングおじさんのニット帽が新しくなってるな、", "&"),
+            chi.think("なんて、どうでもいいことばかりが頭を巡る",
+                "周囲に同じ高校のブレザーは見かけない", "二人組の小学生がランドセルを背に歩いて通り過ぎていくけれど、",
+                "特に$meたちを気に掛ける様子もない"),
+            ei.talk("そうだろう？"),
+            chi.do("彼の声に、一度だけがんばって顔を上げてみる",
+                "両目に溜まった涙がすっと零れ、やっぱり$meは泣いてしまった",
+                "空は一部が茜色に塗られているだけで後は全部グレィの雲が蓋をしている",
+                "そこから雨が落ちてきそうだ"),
+            ei.talk("なあ"),
+            chi.do("自転車を脇に置いた彼はじっとこちらを見て何とか返事をもらおうとしているけれど、",
+                "$meはまだ、決断できずにいた"),
+            ei.talk("別れよう？"),
+            chi.explain("その瞬間というものは気づいたらもうそこに立ってしまっていて、後戻りができない",
+                "そういうものなんだなと、珍しくちょっと甘えたような言い方をした彼の声で理解した"),
+            chi.talk("あの"),
+            chi.talk("あの"),
+            chi.do("互いの声がハモる",
+                "どうしてこんな時だけ同じタイミングなんだろう"),
+            chi.talk("なんで、こうなっちゃったんだろうね"),
+            ei.talk("最初から合わなかっただけ、かもな"),
+            chi.think("彼の言う最初は、いつだろう",
+                "告白したときだろうか", "付き合い始めた時だろうか",
+                "関係がギクシャクしていると感じ始めた頃だろうか",
+                "$meは必死に彼を見ようとするけれど、目線を上げた時にはそっと彼の視線が明後日に向けられてしまう"),
+            chi.talk("ずっと、そう思っていたってこと？"),
+            chi.do("埋まらない空間を埋めるように質問を投げつける",
+                "思い返せばいつも会話はぎこちなく、どちらからともなく間をもたせようと質問を投げていた"),
+            ei.talk("そりゃ最初は好きだったよ",
+                "だから告白したんだ",
+                "けど、いつからかわかんねえけど、何か違うって思うようになって",
+                "最近は$lineも憂鬱だった"),
+            chi.think("それが初めて聞く、彼の本当の言葉だった"),
+            ei.talk("だって付き合って半年にもなるのに未だにキスの一つもしてないんだぜ？",
+                "それって付き合ってるって言えるのかよ？"),
+            chi.talk("え……"),
+            chi.think("キスしたい、と思っていたことがショックだった",
+                "考えなかった訳じゃない",
+                "でもそれが決定的な不満になるとは思ってもなかった",
+                "ただ一緒にいて、帰る",
+                "学生だし、彼はバスケが忙しいし、そんなことなかなか考えられないのだと思い込んでいた"),
+            ei.talk("デートだって夏休みの一回こっきりだ",
+                "そりゃ$meだって忙しかったさ",
+                "けど、何だよ",
+                "付き合ってたらもっとこう、なんかあるだろ？",
+                "それともただの男友だちでしかなかったのかよ？"),
+            chi.talk("違う、けど"),
+            ei.talk("$meはお前の付き合うがよくわかんない",
+                "だからこうなるのも仕方ないんじゃないか？",
+                "わかり合えないなら、別れるしかない"),
+            chi.do("彼の真っ直ぐな目は$meを非難していた"),
+            ei.talk("じゃあ"),
+            ei.do("という、いつも橋で別れる時に口にするのと全く同じ言葉を吐き出して自転車に跨ると、",
+                    "背を向けてペダルを漕ぎ出した"),
+            chi.talk("$eisuke、君……"),
+            ei.do("呼びかけても何も答えずに行ってしまう"),
+            chi.do("左折して橋を渡り、そのままどんどん姿が小さくなる彼を$meはずっと見つめていた"),
+            chi.think("――どうしてこうなっちゃったんだろう"),
+            chi.do("それを考えながら、$meは空を見上げる",
+                    "世界を暗くする雲から落ちてきたのは小さな雪粒だった"),
+            # NOTE:
+            #   最後の帰り道は無言のまま、何故か歩幅が揃っている
+            #   どちらからも「別れよう」が出てくる
+            #   それを聞いて、ああ、やっぱり、と互いに思う
+            #   珍しく英輔は饒舌に語る
+            #   自分の知らない彼。でも最初の頃はこんなだった（あとで読者が気づく）
+            #   最初の約束、にフォーカスさせる
+            )
+
+## scenes
+def sc_rainy(w: World):
+    chi, ei = w.get("chisa"), w.get("eisuke")
+    return w.scene("嵐の日に", "雨の日には相合い傘をするような関係だったのに",
+            w.symbol("　　　　十月"),
+            w.comment("季節は秋", "雨が多い"),
+            w.comment("呼び名は進展しているものの、彼の態度が、気持ちが離れつつある。ギクシャク感"),
+            chi.look("雨が降っている",
+                "冷たい雨だ",
+                "道路脇の側溝に汚くなった落ち葉が溜まっている",
+                "走り抜けていく車は水を跳ね、傘を差して歩く小学生の列に声を上げさせた"),
+            chi.be("そんな姿を眺めながら傘を手に、$meは待っている",
+                "郵便局の軒先から滴る雫がコンクリートの上に水たまりを作っていて、",
+                "それがどんどん大きくなっていくのを見ながら、待っている"),
+            chi.do("つい春先までは一人だったから誰かを待つなんてことはしなくても良かった",
+                "誰かを待つ",
+                "そのじりじりとした気持ちが恋の一部なんだと思っていたけれど、",
+                "何だろう、最近はその考えに違和感が生まれ始めていた"),
+            chi.do("鞄から携帯電話を取り出し、彼の連絡がないか確認する", "あった"),
+            chi.do("先生に呼び出し受けて遅れると書かれていて、", "その言葉に少しほっとしたけれど、",
+                "携帯電話を握ったまま、もやもやとした気持ちと共にはっきりとしない空を見上げた"),
+            chi.think("秋季大会が近いからと彼のバスケ部は練習の日々で、",
+                "昨日はその様子を覗きに行ったりしたのだけれど、$meのことはちらりとも見てくれなかった",
+                "ただバスケットボールにだけ、仲間にだけ、視線を向けていた",
+                "その必死さが素敵だと思った", "でも寂しさも同時にあった"),
+            chi.do("まだ次の連絡はない",
+                "携帯電話に蓄積された彼との少ないメッセージのやり取りを振り返ると、", "&"),
+            chi.do("最初の頃は何気ないおはようやおやすみのメッセージが並んでいるだけ",
+                "彼はバスケの話で、シュートがよく決まったとか、練習試合でどうだったとか",
+                "それを$meが褒めたり、すごいと言ったり、",
+                "それでも時々は彼が好きそうな漫画やアニメの話題を振ってみる",
+                "でも「読んどく」「見てみるわ」が並んだ後には何もない"),
+            chi.think("だからいつも$meの思考は「どうして$meなの？」に辿り着く",
+                "彼から告白してきたのにこれでは何だか一方通行みたいで、",
+                "でもそういう真面目な話をできる雰囲気も今はまだない"),
+            chi.do("もう先に帰った方がいいかな、と思った時だった"),
+            ei.come("小雨を跳ね除けながら自転車を前傾姿勢で漕いできた彼が、目の前でブレーキを掛けて停まった"),
+            ei.talk("悪かった", "畑中説教がなげーんだ"),
+            ei.do("そこに浮かぶ笑顔は変わらない",
+                "ただヘルメットから雨粒が滴っている"),
+            chi.talk("今日は練習いいの？"),
+            chi.do("$meは慌てて傘を差し出す"),
+            ei.talk("休養日だってさ", "試合後だからそういう方針",
+                "練習しすぎても駄目だって今のコーチがさ"),
+            chi.talk("そっか"),
+            chi.do("傘の先から落ちた雨が、彼の肩を濡らしていた", "思い切り高くしてみるけれど身長差はどうしようもないみたいだ",
+                "その差は永遠に埋まらない気がして何だか情けなくなる"),
+            chi.talk("ごめん"),
+            ei.talk("ああ、$meが持つわ"),
+            ei.do("彼は自転車から降りると傘を奪うように受け取り、代わりに持ってくれた",
+                "&"),
+            chi.do("でも見上げる先にある傘の赤い布が随分遠い",
+                "$meは一歩、彼に近づいたけれど、それでも背中か、肩か、今度は自分に雫がかかる", "どこに立っても$meの一部があぶれてしまった"),
+            ei.talk("傘ぱくってくればよかったな", "悪い"),
+            chi.talk("盗むのは駄目だよ", "$meがもっと大きな傘持ってくればよかっただけ", "ごめん"),
+            chi.think("最近ごめん、と口癖のように謝っている", "彼もよく悪いって言うばかりで、",
+                "何だろう、このギクシャクした居心地の悪さは",
+                "楽しい、も、嬉しい、も、最近あまり見かけない"),
+            chi.talk("なんか、ごめん"),
+            ei.talk("いいよ", "それより帰ろうぜ？"),
+            chi.talk("うん"),
+            chi.do("頷いた途端に目から雫が落下した"),
+            ei.talk("おいおい", "なんでそう泣くんだよ", "$me、また何か言ったか？"),
+            chi.talk("ううん。傘が、小さかったから"),
+            chi.do("泣きたい訳じゃないけれど、雨に濡れただけなんて誤魔化しを口にすることもできないくらい、",
+                    "精神が疲れているのが分かった"),
+            ei.talk("ああ、もう分かったよ"),
+            chi.do("そう言うと彼は$meに傘を押し付けて、先に歩いて行ってしまう",
+                    "自転車も彼も、どんどん濡れていく",
+                    "ブレザーの紺色が濃くなり、ズボンには滝のようなシミができていた"),
+            chi.talk("待って！"),
+            chi.do("$meは彼を追いかける"),
+            chi.talk("濡れてる！"),
+            ei.talk("別にいいよ、これくらい", "$chisaが泣くよりマシだから"),
+            chi.do("以前ならその言葉と笑顔に救われていた",
+                    "でも今は、付け合せの野菜みたいにどこかしんなりとして、味気ない"),
+            chi.talk("ごめん"),
+            chi.do("そう言って付いていく"),
+            chi.look("交差点の信号は赤くぼやけて、二人の足を止めようとするけれど、", "&"),
+            chi.do("彼は構わずに自転車に跨ると、呻き声を上げて渡って行ってしまった"),
+            chi.talk("$ln_eisuke君！"),
+            chi.do("$meはその後を追いかけられず、濡れた横断歩道を前に、走ることを諦めた"),
+            # NOTE:
+            #   雨の中、秋になっている
+            #   傘を使って互いに近寄れない、心の距離が離れていっていることを見せる
+            )
+
+def sc_apart(w: World):
+    chi, ei = w.get("chisa"), w.get("eisuke")
+    return w.scene("離れる距離", "立ち止まらなくなった彼",
+            chi.talk("ねえ"),
+            chi.do("彼は振り返らないのを"),
+            chi.talk("ねえってば"),
+            ei.talk("ん？"),
+            chi.talk("最近スマホばかり見てる"),
+            ei.talk("そうか？"),
+            chi.talk("やっぱ変。なんかあった？"),
+            ei.talk("なんもないよ"),
+            chi.talk("キスの一つもしてないから？"),
+            ei.talk("ちげーよ。そういうんじゃない"),
+            chi.talk("だったら何？"),
+            ei.do("じっと"),
+            chi.talk("もうちょっと、話してほしい"),
+            ei.talk("何話せばいいんだよ"),
+            chi.talk("なんでもいい。ゲームでもバスケでも何でも"),
+            ei.talk("お前としても楽しくないだろ？"),
+            chi.talk("会話がしたいんじゃない"),
+            ei.talk("わかんねーな"),
+            chi.think("彼の歩く速度が早いと"),
+            # NOTE:
+            #   別日。秋風が冷たくて。コート
+            #   最初から距離が離れている
+            #   少し大きな声で呼び止めて、バスケにたとえられる
+            #   不和。話が噛み合わない
+            #   なんで、をスルーされる
+            ).omit()
+
+
+## scenes
+def sc_handdistance(w: World):
+    chi, ei = w.get("chisa"), w.get("eisuke")
+    return w.scene("手の届く距離で", "まだ二人は手の届く距離を歩いていた",
+            w.comment("少し季節が動く"),
+            w.symbol("　　　　七月"),
+            chi.come("#歩いている"),
+            ei.come("#歩いている"),
+            chi.look("夏服の白シャツは半袖の袖口が肌に張り付いて、やや心地悪い",
+                "その肌の上を生温い風が抜けていく",
+                "河原では子どもたちが水浴びをしてきゃっきゃと声を上げていて、自転車を引いて歩いている彼が「いいな」と笑う", "&"),
+            chi.do("それを追いかけるように隣で$meも「うん」と相槌する"),
+            chi.do("この一月半ばかりの下校時の日常だった"),
+            chi.do("鞄を彼の自転車の前カゴに放り込み、両手をぶらぶらと手持ち無沙汰で歩く",
+                "彼の部活がない日は揃って下校できるからいつもの三倍は空気が美味しい"),
+            ei.talk("$ln_chisa？"),
+            chi.talk("あ、うん", "ごめん"),
+            chi.do("気づくと彼が二メートルも先に進んでいる",
+                "$meは慌てて追いつく",
+                "でもその追いつくまで待っていてくれるのが嬉しい",
+                "恋は沢山の嬉しいが待っていると知れたのも、彼のお陰だ"),
+            chi.talk("県大会、残念だったね"),
+            ei.talk("まあ、な"),
+            chi.do("ずっと言えなかった言葉を口にできてほっとする",
+                "けれど彼は雲のほとんど浮いてない空を見上げて「うん」と溜息を吐く",
+                "その吐息の中にどんな気持ちが含まれているのかは、決して話してくれない"),
+            ei.talk("それより試験の方が頭痛い"),
+            chi.talk("$ln_eisuke君の場合はそうかも知れないね"),
+            ei.talk("何だよ", "これでも国語はちっとはマシなんだぞ？"),
+            chi.think("前に最高得点六十八点だったと自慢そうに語ったことを思い出し「そうだね」と笑った"),
+            ei.do("笑う","&"),
+            chi.think("それがずっと欲しかったものだった",
+                "彼の笑顔", "日陰にいるみたいな自分を照らしてくれそうな、優しい、歯並びが少し悪い、でもとても好きな笑顔がすぐそこにある"),
+            chi.do("自転車のハンドルを握る彼の手はがっちりとして節がしっかりしている",
+                "それをじっと見ていると彼が「ん？」と目線を向ける"),
+            chi.talk("ううん。なんでもない"),
+            chi.think("先に歩き出して$meはバスケをしている時のパスを受けて高くボールを掲げた時の、敵と味方を見極めているような横顔を思い出した恥ずかしさをそっと隠す"),
+            ei.talk("ああ、あれか", "また$meが先に行っちゃったから"),
+            chi.talk("ううん、それは別にいいって言ったよ？", "$meが合わせればいいんだし"),
+            chi.do("二人の歩幅は違う", "気にせずに歩き出すとすぐに大差がついて$meは置いてけぼりだけれど、",
+                "彼に付いていくと決めたから文句は言わずに歩く", "最初の頃に何度も立ち止まって溜息がちに見るものだから約束したのだ――$meががんばって追いつくから"),
+            ei.talk("いや、でもさ、どう考えても$ln_chisaに無理させてるって最近思って",
+                "バスケやってるとさ、ほんとチームって大事だなって思うんだ"),
+            chi.talk("またバスケの話？"),
+            chi.think("彼の話の大半はバスケだった", "漫画も、見るテレビも、ゲームも、何もかもバスケ",
+                "友だちも男ばかりで、遊びに行くよりもとにかくバスケの練習がしたいと$meにまで言うような、",
+                "一人のバスケ好きな男子高校生が彼だった"),
+            ei.talk("いや、バスケの話ってか、バスケの話じゃないけど、やっぱバスケの話か"),
+            chi.do("考え込むように顎に右手を当てていたけれど、そのまま歩き出し、話を続けた"),
+            ei.talk(
+                "一人上手い奴がいてもさ、バスケって一人じゃできないんだ",
+                "ボールを運んでシュートが決まっても、すぐにそんな点は取り返されちまう",
+                "結局攻めるにも守るにも周りの奴らとの協力が必要になるんだ",
+                "だから二人でも三人でもいいけど、チームっていうのがとにかく大事でさ、",
+                "それって自分のことだけじゃなくて相手がどう動くか、どうパスするのか、どこにポジショニングするのかとか、",
+                "注意深く見て理解してなきゃいけない",
+                "$meと$ln_chisaもそんな関係を目指していけばいいんじゃないかなって思うんだけど……分かる？"),
+            chi.do("バスケのことになると息の継ぎ間がないくらいに一気呵成にまくしたてて喋る",
+                    "そもそもバスケットボール自体あまり詳しくないし、時々分からない用語が出るけれどそれをいちいち尋ねてもいられないしで、",
+                    "いつも彼が伝えたいことは半分も理解してあげられない"),
+            chi.talk("$meたちはチームってことかな？　カップルじゃなくて"),
+            ei.talk("英語はよく分かんねえけど、一緒に帰れる時はなるべく一緒に帰ろうって、そういうことかな"),
+            chi.talk("うん、わかった。ありがとうね"),
+            chi.do("笑顔を作る",
+                    "彼のそれとは比べ物にならない、作り物でしかない",
+                    "それでも彼は満足して「ああ」とにっこりしてくれるから、たぶんこれで間違っていない"),
+            chi.look("河川敷の公園ではサッカーをしている小学生たちが楽しそうな声を上げている",
+                "一人の子が蹴ったボールが川に落ちて、それでも気にせずに別の子が入っていって拾ってあげた",
+                "&"),
+            chi.do("自分ならきっと黙って見ている側だろう",
+                    "けど彼はああやって拾ってあげる側だ",
+                    "そんな違いをよく感じる",
+                    "それは遠くから見ているだけでははっきり分からなかった、$meと彼との距離感なんだと思った"),
+            chi.do("でもその違う他人が互いを補い合うように歩み寄って、チームになる",
+                "きっとそういうことなんだ", "それが$meたちの恋だと、彼は言いたいのかも知れない",
+                "そんな考えに至り、頬が緩む"),
+            chi.talk("今度の休み、買い物とか、行く？"),
+            ei.talk("ほんと！？", "あ、けど次の休みは高村たちと試験勉強で集まるんだったわ。悪い"),
+            chi.talk("ううん", "それじゃあ試験終わったら、また考えよう"),
+            ei.talk("そうだな"),
+            chi.do("彼の笑顔に、胸の閊えが取れた思いだ"),
+            chi.do("でも顔を上げると、お別れポイントが近づいていた"),
+            chi.look("鉄骨で組まれた橋が向こう岸まで伸びている",
+                    "その橋が自転車通学と徒歩通学の境界線で、", "&"),
+            chi.explain("$meと彼を分けているものでもある"),
+            ei.talk("もうすぐだな"),
+            chi.talk("うん"),
+            chi.do("いつも繰り返される、確認のようなそれを、彼はどんな思いで口にしているのだろう", "&"),
+            chi.think("ちら、と表情を伺う",
+                    "ちょっと落ち着きがないように見えるのは別れを惜しんでくれているからだろうか"),
+            ei.talk("な、ちょっと聞きたいんだが"),
+            chi.do("彼は歩くのをやめて$meの方を振り返る",
+                    "その目が真剣な様子で、思わず何が来るのだろうと身構えてしまうけれど、"),
+            ei.talk("好きな食べ物ってあるか"),
+            chi.do("なんてことのない質問だった"),
+            chi.talk("好きは特別ないけど嫌いなら結構ある"),
+            ei.talk("$meはグリンピースが駄目でさ、カレーとかピラフとかオムライスとかさ"),
+            chi.talk("うんうん。わかる"),
+            chi.do("それで彼は満足したのか、何度も頷いてから自転車に跨る"),
+            chi.think("思わず「待って」と口にしたくなった", "もう少しだけ一緒にいないかと提案したい", "でも彼はあの笑顔を$meに向けて、"),
+            ei.talk("じゃあ"),
+            ei.do("軽く手を振り、ペダルを漕いで行ってしまう"),
+            chi.talk("うん", "また明日……ばいばい"),
+            chi.do("彼の自転車が橋を渡っていく",
+                    "距離はどんどん離れ、ついさっきまで感じていた肌の熱はいつの間にかひんやりとした心地になり、胸の中央をすっと風が抜けていく",
+                    "それでも$meはいつも彼が橋を渡り切ってしまうまでじっと見つめ、ある期待をする",
+                    "でも今日も彼は振り返らない",
+                    "まだ心の距離が遠い気がした"),
+            # NOTE:
+            #   手を繋げないまま、歩くが、互いに目を合わせたり、彼の話に適当に頷くのすら楽しい日々
+            #   好きな食べ物の話なんかでも楽しい
+            #   けれど尽く「合わない」
+            )
+
+
+## scenes
+def sc_confession(w: World):
+    chi, ei = w.get("chisa"), w.get("eisuke")
+    return w.scene("告白の鼓動", "最初の瞬間から、その不安は過っていた",
+            w.comment("告白される", "場所は学校を少し離れたところがいい"),
+            w.comment("季節は春から初夏くらいが良い。最終的に冬に別れるから"),
+            w.comment("$eisukeだけは自転車を引っ張っている。通学が異なる"),
+            w.symbol("　　　　五月"),
+            w.comment("身長x0.45が目安"),
+            chi.come("身長百五十一センチの歩幅は約六十八センチらしい", "&"),
+            chi.do("その幅を重ねて$meは高校を出て一キロくらいの場所にある横断歩道に近づく",
+                "温かい風が小学校のフェンス越しの葉桜を揺らしていくのを横目に歩いていく$meを、",
+                "黄色いカバーを付けたランドセルを背負った小さな子たちが集団で追い抜いていく",
+                "あの頃からずっと、大きな声で子ども特有の高い音域で喚くように喋り合う同級生たちを、",
+                "少し離れて羨ましさを噛み殺しながら見ていた",
+                "そんな$meは今でもこうして一人で、家路を歩いている"),
+            chi.do("横断歩道の信号は赤へと変わり、$meと小学生の集団を分断した",
+                "メロディは止まり、デイサービスの送迎車が慌てて黄色信号を渡っていく"),
+            chi.look("鞄をスカートの前にして立ち止まると、紺色の野暮ったいブレザーが陽射しに温められる","&"),
+            chi.think("衣替えはまだ先だからと体操着に着替えて「知るかよ」と友だちと笑っていた同級生の角刈りの男子の笑顔を思い出し、",
+                "$meも前の二つボタンを外してみた", "汗ばむ白シャツにいくらか外気が入り込んで、気持ち、涼しくなった"),
+            chi.think("学校の中ではまともに彼を見ることも、ちょっと真似をしてみることでさえも、全然できない",
+                "それでもこうやってあの笑顔を思い出すだけで元気が湧いた", "$meにとって、彼は、やや息苦しさを感じる学生生活の光明だった"),
+            chi.hear("体育館を覗きに行くといつも大きく荒い息で「パース！」と叫んでいる",
+                "重そうな体を揺すって走ってはあの大きな手でボールを掴み、がむしゃらにゴールに向かっていく",
+                "こうして目を閉じて耳を澄ますだけで、頭の中で彼の息遣いが再生されるのだ"),
+            chi.think("そう思っていた"),
+            ei.come("#そこに$Sが立っていた"),
+            ei.talk("あ、あの！"),
+            chi.do("急ブレーキの耳障りな音に続いた大きな声に振り返るとそこに、$meより三十センチも大きな体操着姿の男性が肩を揺らして立っていた", "自転車から降りて、慌ててスタンドを立てる",
+                "彼は一度下を向き、それから顔を上げてもう一度「あの」と息混じりの声で$meに話しかける",
+                "それが本当に$meなのか分からなくて左右を見て、他に人がいないのを確認したけれど、彼は「$ln_chisaさん」と呼ぶから思わず「はい」と返事をしてしまった"),
+            chi.do("真正面から顔を見たことなんてないから彼の太くて力強い眉が困ったように曲がり、",
+                "それでも真剣な眼差しが自分に向けられていて、一秒くらいは我慢したけれど、思わず顔を背けてしまう",
+                "明らかに気まずい"),
+            ei.talk("いや、その、別に怖い話じゃないんだ"),
+            chi.talk("う、うん"),
+            chi.do("ぎゅっと掴まれたような喉から無理矢理に出した声が、裏返る"),
+            ei.talk("あの、今さ、付き合ってる人とか、その、彼氏とか、いる？"),
+            chi.talk("え？"),
+            chi.do("声にならない掠れた呟きを、$meは何度も自分の胸の内で繰り返す",
+                "彼は短い髪の毛をガリガリと掻き毟り、口なのか鼻からなのか分からない荒い息遣いで答えを待っている",
+                "そのハァハァスゥ、という彼の息のリズムが$meの鼓動を大きく狂わせ、ぎゅっと胸の前に右手を持っていく",
+                "苦しい、という気持ちを辛うじて「ううん」という返事に変える"),
+            ei.talk("え？"),
+            chi.talk("い、いません"),
+            chi.do("はっきりと否定することに唇が震える"),
+            ei.talk("じゃあ、好きな人とかは？"),
+            chi.talk("えっ"),
+            ei.talk("あ、ごめ", "今の無し", "ちょっと待って"),
+            chi.do("答える暇もなく彼が慌ててその大きな逞しい手を広げて質問をかき消そうとするのが何だかおかしくて、",
+                "小さく吹き出した"),
+            ei.talk("あ、だから、そうじゃなくて"),
+            ei.do("彼は咳払いをして深呼吸で整えると、いつも友だちと話しているのとは全然違う低く、けれどしっかりと伝わる声でこう言った"),
+            ei.talk("$meと、付き合って下さい", "お願いします"),
+            ei.do("右手を出し、思い切り頭を下げる"),
+            chi.do("$meは慌ててとりあえず周囲を見た", "&"),
+            chi.look("車は信号の切り替わりで止まり、横断歩道は青に変わり、待っていたおばさんが歩き出す",
+                    "新しい小学生の集団の中の一人の女子は向かい合って何だか気まずそうな高校生二人を見て、何かを察しているのかも知れない",
+                    "自転車通学の知らない生徒の男子が、目線だけ向けて$meたちの脇を抜けていった"),
+            chi.do("顔が熱い", "頬が火照る",
+                    "彼の顔なんてもう見れず、それでもじっと右手を出したまま待っているのだと分かる"),
+            chi.do("どうすればいい？", "どう答えればいい？"),
+            chi.do("その自問が頭の中で繰り返される",
+                    "まるで左脳と右脳が喧嘩を始めたみたいに思考の箱がいっぱいになって、溢れて、",
+                    "$meは何も言えないまま鼻水をすすり始める"),
+            ei.talk("あの、$ln_chisaさん？"),
+            chi.hear("横断歩道のメロディは音が半分外れたような調子でずっと鳴り響いている"),
+            chi.talk("なんで、$me？"),
+            chi.do("やっと絞り出したのはそんな問いかけで、", "彼がどんな表情をしているかなんて見られずに$meは胸の右手を再び鞄の握りに戻すと、",
+                    "俯いたままの瞳から涙を落とした"),
+            ei.talk("なんでって、その、好きだけじゃ駄目、かな"),
+            chi.hear("好き、という言葉の響きは甘酸っぱい", "でもそれは砂糖なのか毒なのか、とても判別できそうにない",
+                    "指先からじわりと痺れが広がって、体温が行方不明だ"),
+            ei.talk("好きだけじゃ、駄目かよ？"),
+            chi.hear("声が強くなる",
+                    "周りの人にもきっと聞こえている"),
+            chi.think("逃げ出したい",
+                    "それなのに「駄目か？」と彼の声が追い打ちをする",
+                    "$meの足にそれぞれ十キロくらいの重石を乗せる"),
+            ei.talk("……泣いて、る？"),
+            chi.do("声は出ず、ただ頷くだけだ"),
+            ei.talk("じゃあ駄目か？"),
+            ei.do("諦めたように呟いて、彼の温度が遠ざかる"),
+            chi.talk("駄目、じゃないよ"),
+            chi.do("顔を上げた", "彼は自転車に跨がり、行ってしまおうとしている",
+                    "でも「じゃないよ」ともう一度言うと顔をこちらに向けてくれた"),
+            ei.talk("あの、泣かせるつもりじゃなかった",
+                    "てか、泣くと思ってなかった",
+                    "突然こんなこと言って、ごめんかった"),
+            chi.talk("ごめん、くない", "ごめんくないよ"),
+            ei.talk("え？"),
+            chi.talk("嬉しいって言ったの", "嬉しいんだよ？"),
+            chi.do("彼は困惑しつつも「じゃあ」と$meの顔を覗き込む"),
+            chi.do("それに小さく頷くと、やっと彼の表情が明るくなった",
+                    "気持ちが通じた、と思える瞬間だった"),
+            ei.talk("そっか。そっかよ"),
+            chi.do("彼は自転車を降りて何度も「そっか」と呟く", "声が上擦って、でもとても嬉しそうで何度か言っているうちに「やった」「やった！」とガッツポーズまで始める"),
+            chi.talk("$ln_eisuke君、ちょっと"),
+            chi.look("信号はまた赤に変わり、目の前を車が走り抜けていった"),
+            chi.talk("危ないよ？"),
+            ei.talk("大丈夫大丈夫", "今の$meは無敵だから"),
+            ei.do("笑顔だ", "&"),
+            chi.do("彼の太陽みたいな笑顔が、すぐそこにあって、自分へと向けられている",
+                    "ずっと欲しかった、優しくて温かいそれが、自分のものになった瞬間が、この時だった"),
+            ei.talk("えっと、これから、宜しくお願いします"),
+            chi.talk("こ、こちらこそ", "お願いします"),
+            w.comment("ある橋まで、で分かれていることを示す"),
+            # NOTE:
+            #   告白の返事をした後の、ぎこちなさから開始
+            #   一緒に帰ろうか、のところ
+            #   まだ春の残り香が街にある
+            #   爽やかな空とかの情景
+            )
+
+def sc_samestep(w: World):
+    chi, ei = w.get("chisa"), w.get("eisuke")
+    return w.scene("歩幅を合わせて",
+            w.comment("季節は２学期が始まりで、秋。まだ残暑"),
+            w.comment("告白を終えて、一緒に初めて帰った時の歩幅。合わせようとしていた。どちらからともなく"),
+            w.br(),
+            chi.come("#歩いている"),
+            ei.come("#前を歩いている"),
+            chi.look("夕陽が河川敷の片方の土手を大きく陰にしてしまう",
+                "川は水量が少なくてぽつりぽつりと離れ小島みたいな岩がいくつか見えて、","&"),
+            chi.do("そこを桜の花びらが流れていく", "前を向くと五メートルくらい彼との距離が離れてしまっていた"),
+            chi.do("待って、とも言えず、小さな歩幅で懸命に追いつく",
+                "彼は自転車を押しているだけなのに随分と早くて、のんびりと歩く$meとはすぐに差がついてしまった"),
+            ei.talk("あ、悪い"),
+            chi.talk("ううん", "気にしないで"),
+            chi.do("彼は肩越しに振り返り、立ち止まって$meを待ってくれる",
+                "百八十の彼の歩幅は約八十一センチ、一歩で十センチ以上も$meを置いていってしまう",
+                "だから$meが早足になって追いかけるしかなくて、",
+                "一緒に帰ろうと言ってくれた彼は面倒な女だとでも思っているかも知れない",
+                "そんなことを考えると、目に涙が溜まる"),
+            chi.talk("あの、さ"),
+            )
+
+def sc_samestep2(w: World):
+    chi, ei = w.get("chisa"), w.get("eisuke")
+    return w.scene("歩幅を合わせて2",
+            ei.talk("ん？"),
+            chi.hear("彼の低い声が、響く"),
+            chi.do("何か話していないと泣いてしまいそうで、けれど話題を探してもまず言葉が浮かんでこなくて、",
+                "目線は下を向き、歩く速度はどんどん遅くなって立ち止まってしまう", "だから"),
+            chi.talk("何でもない"),
+            chi.do("そう言って諦めた"),
+            ei.talk("$ln_chisaさんはさ、やっぱ本とか読んだりするタイプなの？"),
+            chi.do("けれどそんな$meに彼は優しく話しかける"),
+            ei.talk("ほら、いつも教室では大人しくしてるグループじゃん",
+                "$ln_miyoさんとかと一緒で"),
+            chi.talk("う、うん"),
+            chi.do("唯一話しかけてくれるのが前の席の$full_miyoだった",
+                "ただ友だちと呼ぶまでの関係性はまだ築けていない"),
+            ei.talk("$ln_chisaさんはさ、他の女子みたいにうるさくない人だから、やっぱ本かなって思って"),
+            chi.think("質問の意図が呑み込めなかった",
+                "それでも彼の言葉から何を言いたいのか、その気持ちをがんばって汲み取ろうと次の言葉を待った"),
+            ei.talk("$meは教科書もろくに読まないんだけど、漫画だと読む",
+                "って言ってもやっぱバスケ漫画が嬉しいけど",
+                "ただそれだけじゃバカになるからって兄貴がやたらと日本の歴史とか三国志とか貸してくれるんだ",
+                "知ってるかな、三国志"),
+            chi.talk("うん", "有名な中国の歴史ものだよ"),
+            ei.talk("兄貴はやたら三国志が好きみたいでゲームも三国志、漫画も三国志、小説も三国志、おまけに人生も三国志から学べって",
+                "将来三国志学者にでもなればいいんじゃねえかなって本気で思ってる"),
+            chi.do("そう思わない？　という、いつも友だちにして見せるみたいな悪戯坊主の笑みが$meに向けられる", "&"),
+            chi.do("たぶん話題を作ろうとしてくれているのだろう",
+                "その心遣いが少し息苦しい", "けどその苦しさが恋なのかも知れないと、$meは思っていた"),
+            chi.talk("そうだね"),
+            chi.do("$meの返事を待ってから、彼はゆっくり歩き出す",
+                "自転車の車輪のカラカラと回る音が少しずつ自分から離れていく",
+                "あ、と思ってそれに必死で追いついて、でもまた離れて、を繰り返す"),
+            chi.look("いつの間にか河川敷の路を歩く人影は、$meと彼、二つになっていた",
+                "その間を風が抜けていく"),
+            chi.talk("少し涼しくなってきたね"),
+            ei.talk("そうか？"),
+            chi.look("彼は体操着で、もう一人だけ先に夏にいるみたい",
+                    "対して自分はまだ冬服のブレザーが重い", "ボタンを外しているくらいでは、彼に追いつけない"),
+            chi.talk("$ln_eisuke君は、いつも涼しそう"),
+            ei.talk("ああ、コレな", "女子はジャージや体操着って訳にはいかないもんな"),
+            chi.do("彼はシャツを引っ張って$meを見る",
+                    "まださっき告白されたところなのにもうこんなに近くで自分を見てくる",
+                    "好きだから？　恥ずかしくないのかな",
+                    "分からない",
+                    "その照れ臭さを隠すように下を向いて早足になるけれど、彼はすぐに追いついて、行き過ぎたとまた立ち止まる",
+                    "顔を上げるとそこに、彼の笑顔が待っていた"),
+            chi.talk("ねえ、あの、さ"),
+            # NOTE:
+            #   歩幅を合わせようと、千紗が急ぎ足で追いつこうとするところ
+            #   そこで背中を見つめて、思うこと
+            ).omit()
+
+def sc_samestep3(w: World):
+    chi, ei = w.get("chisa"), w.get("eisuke")
+    return w.scene("歩幅を合わせて3",
+            ei.talk("何？"),
+            chi.talk("もしどっちかが駄目ってなったら、その時は後腐れなく、別れるって約束、して下さい"),
+            chi.do("彼は一瞬何を言われたのか分からなかったみたいだったけれど、一分ほど考え込んで、"),
+            ei.talk("分かった"),
+            chi.look("夕日が落ちてくる",
+                "二人の影が長く伸び、歩き出した時に重なっていたそれは、ゆっくりと離れていった"),
+            w.symbol("（了）"),
+            )
+
